Exception classes moved into single file
