--- src/main/scala/play/api/cache/redis/exception/ExecutionFailedException.scala
@@ -1,6 +1,0 @@
-package play.api.cache.redis.exception
-
-/**
-  * @author Karel Cemus
-  */
-class ExecutionFailedException( key: String, command: String, cause: Throwable ) extends RedisException( command, cause )
