--- src/main/scala/play/api/cache/redis/exception/SerializationException.scala
@@ -1,8 +1,0 @@
-package play.api.cache.redis.exception
-
-/**
-  * Value serialization or deserialization failed.
-  *
-  * @param message
-  */
-class SerializationException( value: Any, message: String ) extends RedisException( message )
