--- src/main/scala/play/api/cache/redis/exception/TimeoutException.scala
@@ -1,6 +1,0 @@
-package play.api.cache.redis.exception
-
-/**
-  * @author Karel Cemus
-  */
-class TimeoutException( key: String ) extends RedisException( "" )
