--- src/main/scala/play/api/cache/redis/exception/UnexpectedResponseException.scala
@@ -1,3 +1,0 @@
-package play.api.cache.redis.exception
-
-class UnexpectedResponseException( key: String, command: String ) extends RedisException( command )
